+ Added squirrel module for libcurl
git-svn-id: https://winglib.svn.sourceforge.net/svnroot/winglib/trunk@225 38e6eff2-4521-0410-8f8e-9afd2f83d171

+++ app/sq_modules/sqmod_irrlicht/sqirr_color.h
@@ -7,8 +7,8 @@ class CSqirrColor
 public:
 
 	_SQBIND_CLASS_CTOR_BEGIN( CSqirrColor )
-		_SQBIND_CLASS_CTOR( CSqirrColor, 3 ) ( sa.GetFloat( 2 ), sa.GetFloat( 3 ), sa.GetFloat( 4 ), 255 )
-		_SQBIND_CLASS_CTOR( CSqirrColor, 4 ) ( sa.GetFloat( 2 ), sa.GetFloat( 3 ), sa.GetFloat( 4 ), sa.GetFloat( 5 ) )
+		_SQBIND_CLASS_CTOR( CSqirrColor, 3 ) ( sa.GetInt( 2 ), sa.GetInt( 3 ), sa.GetInt( 4 ), 255 )
+		_SQBIND_CLASS_CTOR( CSqirrColor, 4 ) ( sa.GetInt( 2 ), sa.GetInt( 3 ), sa.GetInt( 4 ), sa.GetInt( 5 ) )
 	_SQBIND_CLASS_CTOR_END( CSqirrColor )
 
 	/// Default constructor
